feature(card): bottom view container refactoring
